+++ src/Resources/assets/js/contao-google-charts-bundle.js
@@ -1,3 +1,0 @@
-import { GoogleChartsBundle } from '../../npm-package';
-
-document.addEventListener('DOMContentLoaded', GoogleChartsBundle.init);