+++ src/Resources/npm-package/js/contao-google-charts-bundle.js
@@ -1,7 +1,0 @@
-
-class GoogleChartsBundle {
-    static init() {
-    }
-}
-
-export {GoogleChartsBundle};